Removed from the instance resource state machine the states related to the master nodes (#791)
