+++ service/controller/resource/instance/create.go
@@ -19,8 +19,6 @@ func (r *Resource) configureStateMachine() {
 		DeploymentInitialized:          r.deploymentInitializedTransition,
 		ProvisioningSuccessful:         r.provisioningSuccessfulTransition,
 		ClusterUpgradeRequirementCheck: r.clusterUpgradeRequirementCheckTransition,
-		MasterInstancesUpgrading:       r.masterInstancesUpgradingTransition,
-		WaitForMastersToBecomeReady:    r.waitForMastersToBecomeReadyTransition,
 		ScaleUpWorkerVMSS:              r.scaleUpWorkerVMSSTransition,
 		CordonOldWorkers:               r.cordonOldWorkersTransition,
 		WaitForWorkersToBecomeReady:    r.waitForWorkersToBecomeReadyTransition,
@@ -33,16 +31,7 @@ func (r *Resource) configureStateMachine() {
 	r.stateMachine = sm
 }
 
-// EnsureCreated operates in 3 different stages which are executed sequentially.
-// The first stage is for uploading ARM templates and is represented by stage
-// DeploymentInitialized.
-// The second stage is for waiting for ARM templates to be applied and is represented
-// by stage ProvisioningSuccessful.
-// The third stage is for draining and upgrading the VMSS instances and is represented by
-// stage InstancesUpgrading.
-// Once all instances are Upgraded the state becomes DeploymentCompleted and the reconciliation
-// loop stops until a change in the ARM template or parameters is detected.
-// Check docs/instances-stages-v13.svg file for a grafical representation of this process.
+// This resource applies the ARM template for the worker instances, monitors the process and handles upgrades.
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	cr, err := key.ToCustomResource(obj)
 	if err != nil {
