+++ service/controller/resource/instance/create_drain_old_worker_nodes.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	corev1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/core/v1alpha1"
+	providerv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
 	"github.com/giantswarm/microerror"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -122,4 +123,46 @@ func (r *Resource) drainOldWorkerNodesTransition(ctx context.Context, obj interf
 	r.logger.LogCtx(ctx, "level", "debug", "message", "deleted all drainerconfigs")
 
 	return TerminateOldWorkerInstances, nil
+}
+
+func (r *Resource) createDrainerConfig(ctx context.Context, customObject providerv1alpha1.AzureConfig, nodeName string) error {
+	r.logger.LogCtx(ctx, "level", "debug", "message", "creating drainer config for tenant cluster node")
+
+	n := key.ClusterID(customObject)
+	c := &corev1alpha1.DrainerConfig{
+		ObjectMeta: metav1.ObjectMeta{
+			Labels: map[string]string{
+				key.ClusterIDLabel: key.ClusterID(customObject),
+			},
+			Name: nodeName,
+		},
+		Spec: corev1alpha1.DrainerConfigSpec{
+			Guest: corev1alpha1.DrainerConfigSpecGuest{
+				Cluster: corev1alpha1.DrainerConfigSpecGuestCluster{
+					API: corev1alpha1.DrainerConfigSpecGuestClusterAPI{
+						Endpoint: key.ClusterAPIEndpoint(customObject),
+					},
+					ID: key.ClusterID(customObject),
+				},
+				Node: corev1alpha1.DrainerConfigSpecGuestNode{
+					Name: nodeName,
+				},
+			},
+			VersionBundle: corev1alpha1.DrainerConfigSpecVersionBundle{
+				Version: "0.2.0",
+			},
+		},
+	}
+
+	_, err := r.g8sClient.CoreV1alpha1().DrainerConfigs(n).Create(c)
+	if errors.IsAlreadyExists(err) {
+		r.logger.LogCtx(ctx, "level", "debug", "message", "did not create drainer config for tenant cluster node")
+		r.logger.LogCtx(ctx, "level", "debug", "message", "drainer config for tenant cluster node does already exist")
+	} else if err != nil {
+		return microerror.Mask(err)
+	} else {
+		r.logger.LogCtx(ctx, "level", "debug", "message", "created drainer config for tenant cluster node")
+	}
+
+	return nil
 }
